% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_docs}
\alias{browse_docs}
\title{View GitHub API documentation}
\usage{
browse_docs(action = c("get", "post"), object = c("milestone", "issue",
  "issue event", "repo labels"))
}
\arguments{
\item{action}{Character string denoting action you wish to complete: "get" (list existing) or "post" (creating new)}

\item{object}{Character string denoting object on wish you want to apply an action. Supports
"milestone", "issue", "issue event"}
}
\value{
Returns URL in non-interactive session or launches browser to docs in interactive session
}
\description{
Opens browser to relevant parts of GitHub API documentation to learn more about field
definitions and formatting. Inspired by similar \code{browse_} functions included in
the \code{usethis} package.
}
\examples{
\dontrun{
browse_docs('get', 'milestone')
}
}
