% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_authentication}
\alias{check_authentication}
\title{Check for valid authentication credentials}
\usage{
check_authentication(ref)
}
\arguments{
\item{ref}{Any repository reference being used. Repository information is stripped out and
only authentication credentials are used to determine the rate limit.}
}
\value{
If valid, returns related GitHub username. Else returns 401 Unauthorized error.
}
\description{
Check for valid authentication credentials
}
\examples{
\dontrun{
experigit <- create_repo_ref('emilyriederer', 'experigit')
check_authentication(experigit)
}
}
