% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_credentials}
\alias{check_credentials}
\title{Check for valid credentials and repo permissions}
\usage{
check_credentials(ref)
}
\arguments{
\item{ref}{Any repository reference being used. Repository information is stripped out and
only authentication credentials are validated.}
}
\value{
Prints GitHub username as determined by credentials (if valid) and repo-level permissions (if any),
else throws 401 Unauthorized error.
}
\description{
Check for valid credentials and repo permissions
}
\examples{
\dontrun{
experigit <- create_repo_ref('emilyriederer', 'experigit')
check_authentication(experigit)
}
}
