% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_internet}
\alias{check_internet}
\title{Check internet connection}
\usage{
check_internet()
}
\value{
Returns TRUE is connected to internet and false otherwise
}
\description{
Basic wrapper around \code{curl::has_internet()}
}
\examples{
\dontrun{
check_internet()
}
}
\seealso{
Other check: \code{\link{check_rate_limit}}
}
\concept{check}
