% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_rate_limit}
\alias{check_rate_limit}
\title{Find requests remaining and reset time}
\usage{
check_rate_limit(ref)
}
\arguments{
\item{ref}{Any repository reference being used. Repository information is stripped out and
only authentication credentials are used to determine the rate limit.}
}
\value{
Informative message on requests remaining and reset time
}
\description{
Source: copied from \code{httr} vignette "Best practices for API packages" by Hadley Wickham
}
\examples{
\dontrun{
experigit <- create_repo_ref('emilyriederer', 'experigit')
check_rate_limit(experigit)
}
}
\seealso{
Other check: 
\code{\link{check_internet}()}
}
\concept{check}
