% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_issues}
\alias{get_issues}
\title{Get issues from GitHub repository}
\usage{
get_issues(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Get issues from GitHub repository
}
\examples{
\dontrun{
tidytracker <- create_repo_ref('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues(tidytracker)
tt_issues <- parse_issues(tt_issues_res)
}
}
\seealso{
Other get: \code{\link{get_issues_events}},
  \code{\link{get_milestones}}

Other issues: \code{\link{get_issues_events}},
  \code{\link{parse_issues_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_task_board}}
}
\concept{get}
\concept{issues}
