% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_issues_events}
\alias{get_issues_events}
\title{Get issues events from GitHub repository}
\usage{
get_issues_events(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Get issues events from GitHub repository
}
\examples{
\dontrun{
tidytracker <- create_repo_ref('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues_events(tidytracker)
tt_issues <- parse_issues_events(tt_issues_res)
}
}
\concept{get, issues, events}
