% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_milestones}
\alias{get_milestones}
\title{Get milestones from GitHub repository}
\usage{
get_milestones(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Get milestones from GitHub repository
}
\examples{
\dontrun{
tidytracker <- create_repo_ref("emilyriederer", "tidytracker")
tt_milestones_res <- get_milestones(tidytracker)
milestones <- parse_milestones(tt_milestones_res)
}
}
\concept{get, milestones}
