% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghtrackr.R
\docType{package}
\name{ghtrackr}
\alias{ghtrackr}
\alias{ghtrackr-package}
\title{\code{ghtrackr} package}
\description{
A quick and easy wrapper for working with GitHub and other project management tools
}
\details{
See the README on
\href{https://github.com/emilyriederer/ghtrackr}{GitHub}
}
