% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issue_events}
\alias{parse_issue_events}
\title{Parse issue events from \code{get_issues_events}}
\usage{
parse_issue_events(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue-event
}
\description{
Parse issue events from \code{get_issues_events}
}
\examples{
\dontrun{
myrepo <- create_repo_ref('emilyriederer', 'myrepo')
events_res <- get_issue_events(myrepo, number = 1)
events <- parse_issue_events(events_res)
}
}
\seealso{
Other parse: \code{\link{parse_issues}},
  \code{\link{parse_milestones}}

Other issues: \code{\link{get_issue_events}},
  \code{\link{get_issues}}, \code{\link{parse_issues}},
  \code{\link{post_issue}}, \code{\link{report_progress}},
  \code{\link{viz_waterfall_issues}}

Other events: \code{\link{get_issue_events}}
}
\concept{events}
\concept{issues}
\concept{parse}
