% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issues}
\alias{parse_issues}
\title{Parse issues overview from \code{get_issues}}
\usage{
parse_issues(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue
}
\description{
Parse issues overview from \code{get_issues}
}
\examples{
\dontrun{
tidytracker <- create_repo_reference('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues(tidytracker)
tt_issues <- parse_issues(tt_issues_res)
}
}
\concept{parse, issues}
