% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issues_events}
\alias{parse_issues_events}
\title{Parse issues events from \code{get_issues_events}}
\usage{
parse_issues_events(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue-event
}
\description{
Parse issues events from \code{get_issues_events}
}
\examples{
\dontrun{
tidytracker <- create_repo_reference('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues(tidytracker)
tt_issues <- parse_issues(tt_issues_res)
}
}
\seealso{
Other parse: \code{\link{parse_issues}},
  \code{\link{parse_milestones}}

Other issues: \code{\link{get_issues_events}},
  \code{\link{get_issues}}, \code{\link{parse_issues}},
  \code{\link{post_issue}}, \code{\link{viz_gantt_closed}},
  \code{\link{viz_task_board}}

Other events: \code{\link{get_issues_events}}
}
\concept{events}
\concept{issues}
\concept{parse}
