% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_repo_labels}
\alias{parse_repo_labels}
\title{Parse labels from \code{get_repo_labels}}
\usage{
parse_repo_labels(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
`tibble` datasets with one record / label
}
\description{
Parse labels from \code{get_repo_labels}
}
\examples{
\dontrun{
labels <- get_repo_labels(my_repo) \%>\% parse_repo_labels()
}
}
\seealso{
Other parse: \code{\link{parse_issue_comments}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}},
  \code{\link{parse_milestones}}

Other labels: \code{\link{get_repo_labels}}
}
\concept{labels}
\concept{parse}
