% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-engine.R
\name{patch_engine}
\alias{patch_engine}
\title{Core code for all PATCH calls}
\usage{
patch_engine(api_endpoint, ref, ...)
}
\arguments{
\item{api_endpoint}{API endpoint}

\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined body parameters}
}
\value{
Content of PATCH request as list
}
\description{
Core code for all PATCH calls
}
\keyword{internal}
