% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-engine-rewrite.R
\name{post_engine}
\alias{post_engine}
\title{Core code for all POST calls}
\usage{
post_engine(api_endpoint, ref, ...)
}
\arguments{
\item{api_endpoint}{API endpoint}

\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined body parameters}
}
\value{
Content of POST request as list
}
\description{
Core code for all POST calls
}
\keyword{internal}
