% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_issue}
\alias{post_issue}
\title{Issue title (required)}
\usage{
post_issue(ref, title, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined body parameters}
}
\value{
Content of POST request as list
}
\description{
Post issue to GitHub repository
}
\examples{
\dontrun{
tidytracker <- create_repo_ref('emilyriederer', 'tidytracker')
post_issue(tidytracker,
  title = 'this is my issue's title',
  body = 'this is my issue's body',
  labels = c('priority:high', 'bug'))
}
}
\concept{post, issues}
