% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projmgr.R
\docType{package}
\name{projmgr}
\alias{projmgr}
\alias{projmgr-package}
\title{\code{projmgr} package}
\description{
A quick and easy wrapper for working with GitHub and other project management tools
}
\details{
See the README on
\href{https://github.com/emilyriederer/projmgr}{GitHub}
}
