% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_plan}
\alias{report_plan}
\title{Print plan in RMarkdown friendly way}
\usage{
report_plan(plan)
}
\arguments{
\item{plan}{List of project plan, as returned by \code{read_plan_todo_yaml()}}
}
\value{
Returns character string of HTML with class attribute to be correctly
    shown "as-is" in RMarkdown
}
\description{
Interprets list representation of plan, using HTML to format results in a highly readable and
attractive way. Resulting object returned is a character vector of HTML code with the added
class of \code{'knit_asis'} so that when included in an RMarkdown document knitting to HTML,
the results will be correctly rendered as HTML.
}
\examples{
\dontrun{
In RMarkdown:
```{r}
my_plan <- read_plan_todo_yaml("my_plan.yaml")
report_plan(my_plan)
```
}
}
\seealso{
Other plans and todos: \code{\link{post_plan}},
  \code{\link{post_todo}},
  \code{\link{read_plan_todo_yaml}},
  \code{\link{report_todo}}, \code{\link{template_yaml}}
}
\concept{plans and todos}
