% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan-todo.R
\name{template_plan_todo_yaml}
\alias{template_plan_todo_yaml}
\title{Print YAML template to console}
\usage{
template_plan_todo_yaml(template = c("plan", "todo"))
}
\arguments{
\item{template}{One of \code{"plan"} or \code{"todo"} denoting template desired}
}
\value{
Prints template to console
}
\description{
Prints YAML templates for either a plan or to-do list to the console as an example
for developing your own custom plans and to-do lists.
}
\examples{
template_plan_todo_yaml('plan')
template_plan_todo_yaml('todo')
}
\seealso{
Other plans and todos: \code{\link{post_plan}},
  \code{\link{post_todo}},
  \code{\link{read_plan_todo_yaml}}
}
\concept{plans and todos}
