% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytracker.R
\docType{package}
\name{tidytracker}
\alias{tidytracker}
\alias{tidytracker-package}
\title{\code{tidytracker} package}
\description{
A quick and easy wrapper for working with GitHub and other project management tools
}
\details{
See the README on
\href{https://github.com/emilyriederer/tidytracker}{GitHub}
}
