% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{update_issue}
\alias{update_issue}
\title{Update issue in github repository}
\usage{
update_issue(ref, issue_number, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{create_repo_ref()}}

\item{issue_number}{Issue number}

\item{...}{Additional user-defined body parameters. Use \code{browse_docs()} to learn more.}
}
\value{
Number (identifier) of updated issue
}
\description{
Update issue in github repository
}
\examples{
\dontrun{
myrepo <- create_repo_ref('emilyriederer', 'myrepo')
update_issue(myrepo,
  issue_number = 1,
  labels = c('priority:high', 'bug'))
}
}
\seealso{
Other issues: 
\code{\link{get_issue_comments}()},
\code{\link{get_issue_events}()},
\code{\link{get_issues}()},
\code{\link{parse_issue_comments}()},
\code{\link{parse_issue_events}()},
\code{\link{parse_issues}()},
\code{\link{post_issue}()},
\code{\link{report_discussion}()},
\code{\link{report_progress}()},
\code{\link{viz_waterfall}()}
}
\concept{issues}
