% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-engine.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate that all user-defined inputs (GET query, POST body, etc.) are valid}
\usage{
validate_inputs(input, allowed_vars)
}
\arguments{
\item{input}{List of user-provided input parameters}

\item{allowed_vars}{Character vector of allowed but not required variables}

\item{required_vars}{Character vector of required variables}
}
\value{
No return. Throws errors if user-defined inputs are invalid.
}
\description{
Validate that all user-defined inputs (GET query, POST body, etc.) are valid
}
\keyword{internal}
