% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-gantt.R
\name{viz_gantt}
\alias{viz_gantt}
\title{Visualize issues by opened/closed date}
\usage{
viz_gantt(issues)
}
\arguments{
\item{issues}{Issues dataset from output of \code{tidytracker::parse_issues}}
}
\value{
Gantt style \code{ggplot2} object with one row / issue
}
\description{
Creates Gannt-style chart showing var for length of time each issue was opened.
Closed issues are colored by duration of time open. Open issues are grey.
}
\examples{
\dontrun{viz_gantt(issues)}
}
