% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-gantt.R
\name{viz_gantt_closed}
\alias{viz_gantt_closed}
\title{Visualized closed issues by open date}
\usage{
viz_gantt_closed(issues)
}
\arguments{
\item{issues}{Issues dataset from output of \code{tidytracker::parse_issues}}
}
\value{
Gantt style \code{ggplot2} object with one row / issue
}
\description{
Creates Gannt-style chart showing var for length of time each issue was opened.
Issues are colored by duration of time open.
}
\examples{
\dontrun{
dplyr <- create_repo_ref('tidyverse', 'dplyr')
dplyr_issues_raw <- get_issues(dplyr, milestone = 1, state = "all")
dplyr_issues <- parse_issues(dplyr_issues_raw)
dplyr_issues \%>\%
 mutate_at(vars(created_at, closed_at), lubridate::as_date ) \%>\%
 viz_gantt_closed()
}
}
