% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-gantt.R
\name{viz_gantt_closed_linkedfile}
\alias{viz_gantt_closed_linkedfile}
\title{Save SVG of Gantt-style chart of closed issues with links to issues}
\usage{
viz_gantt_closed_linkedfile(g, filepath)
}
\arguments{
\item{g}{ggplot2 object returned by \code{viz_gantt_closed()}}

\item{filepath}{Location to save resulting SVG file of ggplot2}
}
\value{
Writes SVG to file and also returns the body so that is can be easily put
    into an RMarkdown (with use of the \code{results = 'asis'}) chunk option
}
\description{
This function creates the same plot as \code{viz_gantt_closed} then edits the
underlying XML so that the titles of the issues on the y-axis are linked to the
actual issue on GitHub
}
\details{
Credit goes to this Stack Overflow answer for figuring out how to do this:
https://stackoverflow.com/questions/42259826/hyperlinking-text-in-a-ggplot2-visualization/42262407
}
\examples{
\dontrun{
# In R:
viz_gantt_closed_linkedfile(issues, "my_folder/my_file")

# In RMarkdown knitting to HTML:
```{r results = 'asis', echo = FALSE}
cat(readLines("my_folder/my_file"), sep = "\\n")
````
}
}
\seealso{
Other issues: \code{\link{get_issues_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_taskboard_linkedfile}},
  \code{\link{viz_taskboard}}
}
\concept{issues}
