% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-gantt.R
\name{viz_gantt_open}
\alias{viz_gantt_open}
\title{Visualized open issues by opened/closed date}
\usage{
viz_gantt_open(issues)
}
\arguments{
\item{issues}{Issues dataset from output of \code{tidytracker::parse_issues}}
}
\value{
Gantt style \code{ggplot2} object with one row / issue
}
\description{
Creates Gannt-style chart showing var for length of time each issue was opened.
Issues are colored by duration of time open to present.
}
\examples{
\dontrun{viz_gantt_open(issues)}
}
