% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-linked.R
\name{viz_linked}
\alias{viz_linked}
\title{Add GitHub links to select visualizations}
\usage{
viz_linked(g, filepath)
}
\arguments{
\item{g}{ggplot2 object returned by \code{viz_gantt_closed()} or \code{viz_taskboard()}}

\item{filepath}{Location to save resulting SVG file of ggplot2, if desired. Leave blank for
function to output message precisely as needed to render in HTML RMarkdown with chunk
option \code{results = 'asis'}}
}
\value{
SVG version of ggplot2 object with links to relevant GitHub issues. Either writes output
    to file or to console (to be captured in RMarkdown) depending on existence of \code{filepath} argument
}
\description{
This function alters the internal representation of a plot to include links back to the actual
GitHub issue. This is currently implemented for \code{viz_taskboard()} and \code{viz_gantt()}
}
\details{
Credit goes to this Stack Overflow answer for figuring out how to do this:
https://stackoverflow.com/questions/42259826/hyperlinking-text-in-a-ggplot2-visualization/42262407
}
\examples{
\dontrun{
# In R, to save to file:
taskboard <- viz_taskboard(issues)
viz_linked(taskboard, "my_folder/my_file.svg")

# In RMarkdown chunk, to print as output:
```{r results = 'asis', echo = FALSE}
gantt <- viz_gantt_closed(issues)
viz_linked(gantt)
````
}
}
