% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-task-board.R
\name{viz_task_board_svg}
\alias{viz_task_board_svg}
\title{Save SVG of Agile-style task board of issue status with links to issues}
\usage{
viz_task_board_svg(issues, filepath)
}
\arguments{
\item{issues}{Issues dataset as produced by \code{parse_issues()}}

\item{filepath}{Location to save resulting SVG file of ggplot2}
}
\value{
Writes SVG to file and also returns the body so that is can be easily put
    into an RMarkdown (with use of the \code{results = 'asis'}) chunk option
}
\description{
This function creates the same plot as \code{viz_task_board} then edits the
underlying XML so that the "cards" are linked to the corresponding issues on GitHub.
It saves a file with the reuslting SVG, which can then be read into an RMarkdown
HTML document as shown in the Examples.
}
\details{
Credit goes to this Stack Overflow answer for figuring out how to do this:
https://stackoverflow.com/questions/42259826/hyperlinking-text-in-a-ggplot2-visualization/42262407
}
\examples{
\dontrun{
# In R:
viz_task_board_svg(issues, "my_folder/my_file")

# In RMarkdown knitting to HTML:
```{r results = 'asis', echo = FALSE}
cat(readLines("my_folder/my_file"), sep = "\\n")
````
}
}
