% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-waterfall.R
\name{viz_waterfall_issues}
\alias{viz_waterfall_issues}
\title{Visualize waterfall of issues opened, closed, and pending over timeframe}
\usage{
viz_waterfall_issues(issues, start_date, end_date, start = created_at,
  end = closed_at)
}
\arguments{
\item{issues}{Issues dataset as produced by \code{parse_issues()}}

\item{start_date}{Character string in 'YYYY-MM-DD' form for first date to be considered}

\item{end_date}{Character string in 'YYYY-MM-DD' form for last date to be considered}

\item{start}{Unquoted variable name denoting issue start date}

\item{end}{Unquoted variable name denoting issue end date}
}
\value{
ggplot object
}
\description{
Visualize waterfall of issues opened, closed, and pending over timeframe
}
\examples{
\dontrun{
viz_progress_waterfall(issues, '2017-01-01', '2017-03-31')
}
}
\seealso{
Other issues: \code{\link{get_issue_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{report_progress}}
}
\concept{issues}
