% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_docs}
\alias{browse_docs}
\title{View GitHub API documentation}
\usage{
browse_docs(action = c("get", "post"), object = c("milestone", "issue",
  "issue event"))
}
\arguments{
\item{action}{Character string denoting action you wish to complete: "get" (list existing) or "post" (creating new)}

\item{object}{Character string denoting object on wish you want to apply an action. Supports
"milestone", "issue", "issue event"}
}
\value{
Returns URL in non-interactive session or launches browser to docs in interactive session
}
\description{
Read relevant parts of documentation for the GitHub API to learn more about
}
\examples{
\dontrun{
browse_docs('get', 'milestone')
}
}
