% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_issues}
\alias{browse_issues}
\title{Browse issues for given GitHub repo}
\usage{
browse_issues(repo_ref, number = "")
}
\arguments{
\item{repo_ref}{Repository reference as created by \code{create_repo_ref()}}

\item{number}{Optional argument of issue number, if opening page for specific issue is desired}
}
\value{
Returns URL in non-interactive session or launches browser to docs in interactive session
}
\description{
Opens browser to GitHub issues for a given repo. Inspired by similar \code{browse_} functions included in
the \code{usethis} package.
}
\examples{
\dontrun{
my_repo <- create_repo_ref("repo_owner", "repo")
browse_issues(my_repo)
}
}
