% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-engine.R
\name{get_engine}
\alias{get_engine}
\title{Core code for all GET calls}
\usage{
get_engine(api_endpoint, ref, ...)
}
\arguments{
\item{api_endpoint}{API endpoint}

\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Core code for all GET calls
}
\keyword{internal}
