% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_issues}
\alias{get_issues}
\title{Get issues from GitHub repository}
\usage{
get_issues(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{create_repo_ref()}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Get issues from GitHub repository
}
\examples{
\dontrun{
myrepo <- create_repo_ref('emilyriederer', 'myrepo')
issues_res <- get_issues(myrepo)
issues <- parse_issues(issues_res)
}
}
\seealso{
Other get: \code{\link{get_issue_events}},
  \code{\link{get_milestones}}

Other issues: \code{\link{get_issue_events}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{report_progress}},
  \code{\link{viz_gantt_closed_links}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_taskboard_links}},
  \code{\link{viz_taskboard}},
  \code{\link{viz_waterfall_issues}}
}
\concept{get}
\concept{issues}
