% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_issues_events}
\alias{get_issues_events}
\title{Get events for a specific issue from GitHub repository}
\usage{
get_issues_events(ref, number)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{number}{Number of issue}
}
\value{
Content of GET request as list
}
\description{
Get events for a specific issue from GitHub repository
}
\examples{
\dontrun{
tidytracker <- create_repo_ref('emilyriederer', 'tidytracker')
tt_events_res <- get_issue_events(tidytracker, number = 1)
tt_events <- parse_issue_events(tt_events_res)
}
}
\seealso{
Other get: \code{\link{get_issues}},
  \code{\link{get_milestones}}

Other issues: \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{report_progress}},
  \code{\link{viz_gantt_closed_links}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_taskboard_links}},
  \code{\link{viz_taskboard}}

Other events: \code{\link{parse_issue_events}}
}
\concept{events}
\concept{get}
\concept{issues}
