% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_milestones}
\alias{get_milestones}
\title{Get milestones from GitHub repository}
\usage{
get_milestones(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{create_repo_ref()}}

\item{...}{Additional user-defined query parameters}
}
\value{
Content of GET request as list
}
\description{
Get milestones from GitHub repository
}
\examples{
\dontrun{
myrepo <- create_repo_ref("emilyriederer", "myrepo")
milestones_res <- get_milestones(myrepo)
milestones <- parse_milestones(milestones_res)
}
}
\seealso{
Other get: \code{\link{get_issue_events}},
  \code{\link{get_issues}}

Other milestones: \code{\link{parse_milestones}},
  \code{\link{post_milestone}}
}
\concept{get}
\concept{milestones}
