% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_repo_labels}
\alias{get_repo_labels}
\title{Get all labels for a repository}
\usage{
get_repo_labels(ref, ...)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{create_repo_ref()}}

\item{...}{Additional arguments (none currently implemented on GitHub API)}
}
\value{
Content of GET request as list
}
\description{
Get all labels for a repository
}
\examples{
\dontrun{
labels <- get_repo_labels(my_repo) \%>\% parse_repo_labels()
}
}
\seealso{
Other labels: \code{\link{parse_repo_labels}}
}
\concept{labels}
