% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issues}
\alias{parse_issues}
\title{Parse issues overview from \code{get_issues}}
\usage{
parse_issues(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue
}
\description{
Parse issues overview from \code{get_issues}
}
\examples{
\dontrun{
myrepo <- create_repo_reference('emilyriederer', 'myrepo')
issues_res <- get_issues(myrepo)
issues <- parse_issues(issues_res)
}
}
\seealso{
Other issues: \code{\link{get_issue_comments}},
  \code{\link{get_issue_events}}, \code{\link{get_issues}},
  \code{\link{parse_issue_comments}},
  \code{\link{parse_issue_events}},
  \code{\link{post_issue}},
  \code{\link{report_discussion}},
  \code{\link{report_progress}},
  \code{\link{viz_waterfall}}
}
\concept{issues}
