% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issues}
\alias{parse_issues}
\title{Parse issues overview from \code{get_issues}}
\usage{
parse_issues(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue
}
\description{
Parse issues overview from \code{get_issues}
}
\examples{
\dontrun{
tidytracker <- create_repo_reference('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues(tidytracker)
tt_issues <- parse_issues(tt_issues_res)
}
}
\seealso{
Other parse: \code{\link{parse_issue_events}},
  \code{\link{parse_milestones}}

Other issues: \code{\link{get_issues_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{post_issue}}, \code{\link{report_progress}},
  \code{\link{viz_gantt_closed_links}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_taskboard_links}},
  \code{\link{viz_taskboard}}
}
\concept{issues}
\concept{parse}
