% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_issues_events}
\alias{parse_issues_events}
\title{Parse issues events from \code{get_issues_events}}
\usage{
parse_issues_events(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
\code{tibble} datasets with one record / issue-event
}
\description{
Parse issues events from \code{get_issues_events}
}
\examples{
\dontrun{
tidytracker <- create_repo_reference('emilyriederer', 'tidytracker')
tt_issues_res <- get_issues(tidytracker)
tt_issues <- parse_issues(tt_issues_res)
}
}
\concept{parse, issues, events}
