% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_milestones}
\alias{parse_milestones}
\title{Parse milestones from \code{get_milestones}}
\usage{
parse_milestones(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
`tibble` datasets with one record / milestone
}
\description{
Parse milestones from \code{get_milestones}
}
\examples{
\dontrun{
tidytracker_ref <- create_repo_ref("emilyriederer", "tidytracker")
tt_milestones_res <- get_milestones(tidytracker_ref)
tt_milestones <- parse_milestones(tt_milestones_res)
}
}
\seealso{
Other parse: \code{\link{parse_issues_events}},
  \code{\link{parse_issues}}

Other milestones: \code{\link{get_milestones}},
  \code{\link{post_milestone}}
}
\concept{milestones}
\concept{parse}
