% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_milestones}
\alias{parse_milestones}
\title{Parse milestones from \code{get_milestones}}
\usage{
parse_milestones(res)
}
\arguments{
\item{res}{List returned by corresponding \code{get_} function}
}
\value{
`tibble` datasets with one record / milestone
}
\description{
Parse milestones from \code{get_milestones}
}
\examples{
\dontrun{
myrepo <- create_repo_ref("emilyriederer", "myrepo")
milestones_res <- get_milestones(myrepo)
milestones <- parse_milestones(milestones_res)
}
}
\seealso{
Other parse: \code{\link{parse_issue_comments}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}},
  \code{\link{parse_repo_labels}}

Other milestones: \code{\link{get_milestones}},
  \code{\link{post_milestone}}
}
\concept{milestones}
\concept{parse}
