% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{post_plan}
\alias{post_plan}
\title{Post plan (milestones + issues) to GitHub repository}
\usage{
post_plan(ref, plan)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{plan}{Plan list as read with \code{tidytracker::read_plan_yaml}}
}
\value{
Content of POST request as list
}
\description{
Post plan (milestones + issues) to GitHub repository
}
\seealso{
Other plan: \code{\link{read_plan_yaml}}
}
\concept{plan}
