% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_todo.R
\name{post_plan}
\alias{post_plan}
\title{Post plan (milestones + issues) to GitHub repository}
\usage{
post_plan(ref, plan)
}
\arguments{
\item{ref}{Repository reference (list) created by \code{tidytracker::create_repo_ref}}

\item{plan}{Plan list as read with \code{tidytracker::read_plan_yaml}}
}
\value{
Content of POST request as list
}
\description{
Post plan (milestones + issues) to GitHub repository
}
\examples{
\dontrun{
# This example uses example file included in pkg. You should be able to run example as-is
file_path <- system.file("extdata", "plan_yaml.txt", package = "tidytracker", mustWork = TRUE)
my_plan <- read_plan_yaml("ext/my_project_plan.yaml")
post_plan(my_plan)
}
}
\seealso{
Other plan: \code{\link{read_yaml}}
}
\concept{plan}
