% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{read_plan_yaml}
\alias{read_plan_yaml}
\title{Read plan from YAML}
\usage{
read_plan_yaml(filepath = NA, chars = NA)
}
\arguments{
\item{filepath}{Filepath of YAML file. Either this or \code{chars} is required.}

\item{chars}{Character object containing YAML. Either this or \code{chars} in required.}
}
\value{
\code{tibble} containing plan compatible with \code{tidytracker::post_plan}
}
\description{
This function reads a carefully constructed YAML file representing a project plan of
milestones and issues. YAML is converted into a \code{tibble} data structure which can
be passed to \code{tidytracker::post_plan} to buidl repository infrastrucuture.
}
\details{
Please see the "Building Custom Plans" vignette for more details.s
}
\examples{
\dontrun{
plan_tbl <- read_plan_yaml("ext/my_project_plan.yaml")
post_plan(plan_tbl)
}
}
\seealso{
Other plan: \code{\link{post_plan}}
}
\concept{plan}
