% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_progress}
\alias{report_progress}
\title{Print issue-milestone progress in RMarkdown friendly way}
\usage{
report_progress(issues)
}
\arguments{
\item{issues}{Dataframe or tibble of issues and milestones, as returned by
\code{get_issues()} and \code{parse_issues()}}
}
\value{
Returns character string of HTML with class attribute to be correctly
    shown "as-is" in RMarkdown
}
\description{
Interprets dataframe or tibble of issues by breaking apart milestones and listing each
issue title as open or closed, and uses HTML to format results in a highly readable and
attractive way. Resulting object returned is a character vector of HTML code with the added
class of \code{'knit_asis'} so that when included in an RMarkdown document knitting to HTML,
the results will be correctly rendered as HTML.
}
\examples{
\dontrun{
In RMarkdown:
```{r}
issues <- get_issues(repo, state = 'all') \%>\% parse_issues()
report_progress(issues)
```
}
}
\seealso{
Other issues: \code{\link{get_issue_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{viz_waterfall_issues}}
}
\concept{issues}
