% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-gantt.R
\name{viz_gantt_closed}
\alias{viz_gantt_closed}
\title{Visualize Gannt-style chart of closed issues}
\usage{
viz_gantt_closed(issues, start = created_at, end = closed_at)
}
\arguments{
\item{issues}{Issues dataset as produced by \code{parse_issues()}}

\item{start}{Unquoted variable name denoting issue start date}

\item{end}{Unquoted variable name denoting issue end date}
}
\value{
ggplot object
}
\description{
Produces plot of all closed issues from start date to end date, based on output of
issues from \code{tidytracker::get_issues} and \code{tidytracker::parse_issues}. Plot
has one row per issues spanning from \code{start} to \code{end}.
}
\details{
By default, the start date is the issue's \code{created_at} date, and the end date is
the issue's \code{closed_at} date. However, either of these can be altered via the
\code{start} and \code{end} parameters since these dates might not be reflective of the
true timeframe (e.g. if issues are posted well in advance of work beginning.)
}
\examples{
\dontrun{
issues <- get_issues(myrepo, state = "closed") \%>\% parse_issues()
viz_gantt_closed(issues)
}
}
\seealso{
Other issues: \code{\link{get_issues_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{report_progress}},
  \code{\link{viz_gantt_closed_links}},
  \code{\link{viz_taskboard_links}},
  \code{\link{viz_taskboard}},
  \code{\link{viz_waterfall_issues}}
}
\concept{issues}
