% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-taskboard.R
\name{viz_taskboard}
\alias{viz_taskboard}
\title{Visualize Agile-style task board of issue status}
\usage{
viz_taskboard(issues)
}
\arguments{
\item{issues}{Issues dataset as produced by \code{parse_issues()}}
}
\value{
ggplot object
}
\description{
Produces three column task board of all issues categorized as "Not Started", "In Progress",
or "Done". When interativitity is enabled (by changing the  \code{.interactive} parameter
from its default of \code{FALSE}), a tooltip is enabled to show the \code{created_at} date
for each issue.
}
\details{
The following logic is used to determine the status of each issue:
\itemize{
  \item{Not Started:}{ \code{closed_at} is NA and no \code{in-progress} label}
  \item{In Progress:}{ \code{closed_at} is NA but \code{in-progress} label exists}
  \item{Done:}{ \code{closed_at} is not NA}
}
}
\examples{
\dontrun{
issues <- get_issues(myrepo, milestone = 1) \%>\% parse_issues()
viz_task_board(issues)
}
}
\seealso{
Other issues: \code{\link{get_issues_events}},
  \code{\link{get_issues}},
  \code{\link{parse_issue_events}},
  \code{\link{parse_issues}}, \code{\link{post_issue}},
  \code{\link{report_progress}},
  \code{\link{viz_gantt_closed_links}},
  \code{\link{viz_gantt_closed}},
  \code{\link{viz_taskboard_links}},
  \code{\link{viz_waterfall_issues}}
}
\concept{issues}
